requirements.txt
# ── SAP O2C Graph System — Python Dependencies ───────────────────────────────
#
# Install: pip install -r requirements.txt
#
# Phase 1 — Ingestion
pandas==2.2.2

# Phase 2 — Database
psycopg2-binary==2.9.9
python-dotenv==1.0.1

# Phase 3 — LLM Query Engine
fastapi==0.111.0
uvicorn[standard]==0.30.1
langchain==0.2.5
langchain-groq==0.1.5
langchain-community==0.2.5
groq==0.9.0
pydantic==2.7.4

# Phase 4 — Utilities
httpx==0.27.0
python-multipart==0.0.9
